Split Inspector into its own Panel
